container,website: logs via websocket
