+++ bin/container/index.js
@@ -13,6 +13,7 @@ const resource = {
     plugins: genericDefaults.plugins,
     earlyAdoptersOnly: true,
     title: 'container',
+    extraCommands: ['log', 'start', 'stop'],
 };
 
 const category = genericResource(resource);
@@ -33,13 +34,10 @@ const childDefault = Object.assign({}, resource, {
 
 category.addChild(require('./attach')(childDefault));
 category.addChild(require('./process')(childDefault));
-category.addChild(require('./log')(childDefault));
 const actionDefault = Object.assign({}, resource, childDefault, {
     dirname: `${__dirname}/action`,
 });
 
-category.addChild(genericAction(actionDefault, 'stop'));
-category.addChild(genericAction(actionDefault, 'start'));
 category.addChild(genericAction(actionDefault, 'restart'));
 
 module.exports = category;
