--- bin/container/log/examples.md
@@ -1,3 +1,0 @@
-```
-{{command_name}} --agent my-container
-```
