+++ bin/generic/log/examples.md
@@ -1,0 +1,3 @@
+```
+{{command_name}} --{{type}} my-{{type}}
+```
