+++ lib/api.js
@@ -264,14 +264,14 @@ const update = async (method, uri, body) => {
     return res.body;
 };
 
-const getWS = (urn = '', wsUrl = WS_URL) => new Promise((resolve, reject) => {
+const getWS = (urn = '', wsUrl = WS_URL, options={}) => new Promise((resolve, reject) => {
     if (verbose) {
         logger('verbose', `ws ${wsUrl}${urn}`);
     }
 
-    const ws = new WebSocket(`${wsUrl}${urn}`, {
+    const ws = new WebSocket(`${wsUrl}${urn}`, Object.assign({
         headers: getHeaders(),
-    });
+    }, options));
 
     if (!ws) {
         logger('error', 'did not managed to create websocket');
@@ -295,5 +295,8 @@ const getWS = (urn = '', wsUrl = WS_URL) => new Promise((resolve, reject) => {
 
 exports.wsUpload = urn => getWS(`/${urn}`, WS_URL.replace('//api.', '//download.'));
 exports.wsMetrics = urn => getWS(`/v1/${urn}`, WS_URL.replace('/ws', ''));
+exports.wsLogs = urn => getWS(`/v1/${urn}`, WS_URL.replace('/ws', ''), {
+    followRedirects: true,
+});
 exports.getWS = getWS;
 exports.api = api;
